fix: isIntegrationDeleted during prop change - resets the isIntegrationDeleted state when InstallIntegration props change - refactors useIsInstallationDeleted hook inside InstallIntegrationContextProvider - updates import paths
deletes files being moved
