+++ src/components/Configure/content/InstallationContent.tsx
@@ -1,4 +1,6 @@
-import { useInstallIntegrationProps } from 'context/InstallIntegrationContextProvider';
+import {
+  useInstallIntegrationProps,
+} from 'context/InstallIIntegrationContextProvider/InstallIntegrationContextProvider';
 
 import { ErrorTextBox } from '../../ErrorTextBox/ErrorTextBox';
