+++ src/components/Configure/content/fields/ValueMapping/ValuesMapping.tsx
@@ -3,8 +3,10 @@ import {
 } from 'react';
 
 import { ErrorBoundary, useErrorState } from 'context/ErrorContextProvider';
+import {
+  useInstallIntegrationProps,
+} from 'context/InstallIIntegrationContextProvider/InstallIntegrationContextProvider';
 import { FormControl } from 'src/components/form/FormControl';
-import { useInstallIntegrationProps } from 'src/context/InstallIntegrationContextProvider';
 
 import { useSelectedConfigureState } from '../../useSelectedConfigureState';
