+++ src/components/Configure/content/useMutateInstallation.tsx
@@ -3,7 +3,9 @@ import { useCallback } from 'react';
 import { useApiKey } from 'context/ApiKeyContextProvider';
 import { useConnections } from 'context/ConnectionsContextProvider';
 import { ErrorBoundary, useErrorState } from 'context/ErrorContextProvider';
-import { useInstallIntegrationProps } from 'context/InstallIntegrationContextProvider';
+import {
+  useInstallIntegrationProps,
+} from 'context/InstallIIntegrationContextProvider/InstallIntegrationContextProvider';
 import { useProject } from 'context/ProjectContextProvider';
 
 import { useNextIncompleteTabIndex, useSelectedObjectName } from '../nav/ObjectManagementNav/ObjectManagementNavContext';
