+++ src/components/Configure/nav/ObjectManagementNav/v2/index.tsx
@@ -4,7 +4,7 @@ import {
 
 import { Box } from 'components/ui-base/Box/Box';
 import { Container } from 'components/ui-base/Container/Container';
-import { useInstallIntegrationProps } from 'context/InstallIntegrationContextProvider';
+import { useInstallIntegrationProps } from 'context/InstallIIntegrationContextProvider/InstallIntegrationContextProvider';
 import { useProject } from 'context/ProjectContextProvider';
 import { VerticalTabs } from 'src/components/Configure/nav/ObjectManagementNav/v2/Tabs';
 import { NavObject } from 'src/components/Configure/types';
